PhD_Journal/requirements.txt
click==6.7
idna==2.6
ipython-genutils==0.2.0
jsonschema==2.6.0
jupyter-core==4.4.0
MarkupSafe==1.0
nbformat==4.4.0
numpy==1.14.1
pandas==0.22.0
plotly==2.4.1
python-dateutil==2.6.1
pytz==2018.3
scikit-learn==0.19.1
scipy==1.0.0
six==1.11.0
sklearn==0.0
traitlets==4.3.2
